[REFACTORING]: moved VitalityTerms into core-logic
