+++ v6y-apps/front/src/app/layout.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import { ReactNode, Suspense } from 'react';
 
 import VitalityPageLayout from '../commons/components/layout/VitalityPageLayout';
-import VitalityTerms from '../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import VitalityTheme from '../commons/config/VitalityTheme';
 import { AppProvider } from '../infrastructure/providers/AppProvider';
