+++ v6y-apps/front/src/commons/components/application-info/VitalityAppInfos.tsx
@@ -1,4 +1,4 @@
-import VitalityLinks from '@v6y/shared-ui/src/components/VitalityLinks';
+import VitalityLinks from '@v6y/shared-ui/src/components/VitalityLinks/VitalityLinks';
 import { VitalityText } from '@v6y/shared-ui';
 import { Col, Divider, List, Row, Tag } from 'antd';
 import Link from 'next/link';
@@ -7,7 +7,7 @@ import * as React from 'react';
 import useNavigationAdapter from '../../../infrastructure/adapters/navigation/useNavigationAdapter';
 import { QUALITY_METRIC_STATUS } from '../../config/VitalityCommonConfig';
 import VitalityNavigationPaths from '../../config/VitalityNavigationPaths';
-import VitalityTerms from '../../config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import { VitalityAppInfosProps } from '../../types/VitalityAppInfosProps';
 
 const VitalityAppInfos = ({ app, source, canOpenDetails = true, style }: VitalityAppInfosProps) => {
