+++ v6y-apps/front/src/commons/components/help/VitalityHelpView.tsx
@@ -1,7 +1,7 @@
 import { Descriptions } from 'antd';
 import * as React from 'react';
 
-import VitalityTerms from '../../config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import { VitalityModuleType } from '../../types/VitalityModulesProps';
 
 const VitalityHelpView = ({ module }: { module: VitalityModuleType }) => {
