+++ v6y-apps/front/src/commons/components/modules/VitalityModulesView.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { useEffect, useState } from 'react';
 
 import { QUALITY_METRIC_ICONS, QUALITY_METRIC_STATUS } from '../../config/VitalityCommonConfig';
-import VitalityTerms from '../../config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import { VitalityModuleType, VitalityModulesProps } from '../../types/VitalityModulesProps';
 import VitalityDynamicLoader from '../VitalityDynamicLoader';
 import VitalityModal from '../VitalityModal';
