+++ v6y-apps/front/src/features/app-details/components/audit-reports/auditors/lighthouse/VitalityLighthouseReportsCategoryGrouper.tsx
@@ -14,7 +14,7 @@ import {
     QUALITY_METRIC_ICONS,
     QUALITY_METRIC_STATUS,
 } from '../../../../../../commons/config/VitalityCommonConfig';
-import VitalityTerms from '../../../../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import useDataGrouper from '../../../../../../commons/hooks/useDataGrouper';
 import { VitalityModuleType } from '../../../../../../commons/types/VitalityModulesProps';
