+++ v6y-apps/front/src/features/app-details/components/dependencies/VitalityDependenciesView.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import VitalityDynamicLoader from '../../../../commons/components/VitalityDynamicLoader';
 import VitalitySectionView from '../../../../commons/components/VitalitySectionView';
 import VitalityApiConfig from '../../../../commons/config/VitalityApiConfig';
-import VitalityTerms from '../../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import { exportAppDependenciesToCSV } from '../../../../commons/utils/VitalityDataExportUtils';
 import {
     buildClientQuery,
