+++ v6y-apps/front/src/features/app-list/components/VitalityAppListHeader.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import * as React from 'react';
 
 import VitalityApiConfig from '../../../commons/config/VitalityApiConfig';
-import VitalityTerms from '../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import {
     buildClientQuery,
     useClientQuery,
