+++ v6y-apps/front/src/features/apps-stats/components/VitalityAppsStatsChart.tsx
@@ -6,7 +6,7 @@ import * as React from 'react';
 import { useEffect } from 'react';
 
 import VitalityApiConfig from '../../../commons/config/VitalityApiConfig';
-import VitalityTerms from '../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import {
     buildClientQuery,
     useClientQuery,
