+++ v6y-apps/front/src/features/dashboard/components/VitalityDashboardMenu.tsx
@@ -3,7 +3,7 @@ import { Col, Row } from 'antd';
 import * as React from 'react';
 
 import { DashboardItemType } from '../../../commons/config/VitalityCommonConfig';
-import VitalityTerms from '../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import VitalityDashboardMenuItem from './VitalityDashboardMenuItem';
 
 const VitalityDashboardMenu = ({ options }: { options: DashboardItemType[] }) => {
