+++ v6y-apps/front/src/features/dashboard/components/VitalityDashboardView.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 
 import VitalitySearchBar from '../../../commons/components/VitalitySearchBar';
 import { VITALITY_DASHBOARD_DATASOURCE } from '../../../commons/config/VitalityCommonConfig';
-import VitalityTerms from '../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import VitalityDashboardMenu from './VitalityDashboardMenu';
 
 const VitalityDashboardView = () => (
