+++ v6y-apps/front/src/features/notifications/components/VitalityNotificationView.tsx
@@ -6,7 +6,7 @@ import VitalityEmptyView from '@v6y/shared-ui/src/components/VitalityEmptyView/V
 
 import VitalitySectionView from '../../../commons/components/VitalitySectionView';
 import VitalityApiConfig from '../../../commons/config/VitalityApiConfig';
-import VitalityTerms from '../../../commons/config/VitalityTerms';
+import VitalityTerms from '@v6y/core-logic/src/config/VitalityTerms';
 import {
     buildClientQuery,
     useClientQuery,
