+++ v6y-libs/core-logic/src/config/VitalityTerms.ts
