+++ v6y-libs/shared-ui/package.json
@@ -51,6 +51,9 @@
     "eslint-plugin-react-hooks": "=5.1.0",
     "eslint-plugin-react-refresh": "=0.4.18",
     "globals": "=15.14.0",
+    "next": "=15.1.4",
+    "next-compose-plugins": "=2.2.1",
+    "next-i18next": "=15.4.1",
     "prettier": "=3.4.2",
     "typescript": "=5.7.3",
     "typescript-eslint": "=8.20.0",
