+++ v6y-libs/shared-ui/src/types/VitalityLinksProps.ts
@@ -1,0 +1,7 @@
+import { LinkType } from "../../../core-logic/src";
+
+
+export interface VitalityLinksProps {
+    links: LinkType[];
+    align?: 'start' | 'center' | 'end';
+}
